+++ app/src/main/java/com/example/cryptocompose/data/model/gecko/PriceChange24hInCurrency.kt
@@ -7,13 +7,13 @@ data class PriceChange24hInCurrency(
     @SerializedName("aed")
     val aed: Double,
     @SerializedName("ars")
-    val ars: Int,
+    val ars: Double,
     @SerializedName("aud")
     val aud: Double,
     @SerializedName("bch")
     val bch: Double,
     @SerializedName("bdt")
-    val bdt: Int,
+    val bdt: Double,
     @SerializedName("bhd")
     val bhd: Double,
     @SerializedName("bits")
@@ -25,17 +25,17 @@ data class PriceChange24hInCurrency(
     @SerializedName("brl")
     val brl: Double,
     @SerializedName("btc")
-    val btc: Int,
+    val btc: Double,
     @SerializedName("cad")
     val cad: Double,
     @SerializedName("chf")
     val chf: Double,
     @SerializedName("clp")
-    val clp: Int,
+    val clp: Double,
     @SerializedName("cny")
     val cny: Double,
     @SerializedName("czk")
-    val czk: Int,
+    val czk: Double,
     @SerializedName("dkk")
     val dkk: Double,
     @SerializedName("dot")
@@ -53,45 +53,45 @@ data class PriceChange24hInCurrency(
     @SerializedName("hkd")
     val hkd: Double,
     @SerializedName("huf")
-    val huf: Int,
+    val huf: Double,
     @SerializedName("idr")
-    val idr: Int,
+    val idr: Double,
     @SerializedName("ils")
     val ils: Double,
     @SerializedName("inr")
-    val inr: Int,
+    val inr: Double,
     @SerializedName("jpy")
-    val jpy: Int,
+    val jpy: Double,
     @SerializedName("krw")
-    val krw: Int,
+    val krw: Double,
     @SerializedName("kwd")
     val kwd: Double,
     @SerializedName("link")
     val link: Double,
     @SerializedName("lkr")
-    val lkr: Int,
+    val lkr: Double,
     @SerializedName("ltc")
     val ltc: Double,
     @SerializedName("mmk")
-    val mmk: Int,
+    val mmk: Double,
     @SerializedName("mxn")
-    val mxn: Int,
+    val mxn: Double,
     @SerializedName("myr")
     val myr: Double,
     @SerializedName("ngn")
-    val ngn: Int,
+    val ngn: Double,
     @SerializedName("nok")
     val nok: Double,
     @SerializedName("nzd")
     val nzd: Double,
     @SerializedName("php")
-    val php: Int,
+    val php: Double,
     @SerializedName("pkr")
-    val pkr: Int,
+    val pkr: Double,
     @SerializedName("pln")
     val pln: Double,
     @SerializedName("rub")
-    val rub: Int,
+    val rub: Double,
     @SerializedName("sar")
     val sar: Double,
     @SerializedName("sats")
@@ -101,19 +101,19 @@ data class PriceChange24hInCurrency(
     @SerializedName("sgd")
     val sgd: Double,
     @SerializedName("thb")
-    val thb: Int,
+    val thb: Double,
     @SerializedName("try")
-    val tryX: Int,
+    val tryX: Double,
     @SerializedName("twd")
-    val twd: Int,
+    val twd: Double,
     @SerializedName("uah")
-    val uah: Int,
+    val uah: Double,
     @SerializedName("usd")
     val usd: Double,
     @SerializedName("vef")
     val vef: Double,
     @SerializedName("vnd")
-    val vnd: Int,
+    val vnd: Double,
     @SerializedName("xag")
     val xag: Double,
     @SerializedName("xau")
@@ -121,11 +121,11 @@ data class PriceChange24hInCurrency(
     @SerializedName("xdr")
     val xdr: Double,
     @SerializedName("xlm")
-    val xlm: Int,
+    val xlm: Double,
     @SerializedName("xrp")
-    val xrp: Int,
+    val xrp: Double,
     @SerializedName("yfi")
     val yfi: Double,
     @SerializedName("zar")
-    val zar: Int
+    val zar: Double
 )
