+++ app/src/main/java/com/example/cryptocompose/data/model/gecko/PriceChangePercentage14dInCurrency.kt
@@ -25,7 +25,7 @@ data class PriceChangePercentage14dInCurrency(
     @SerializedName("brl")
     val brl: Double,
     @SerializedName("btc")
-    val btc: Int,
+    val btc: Double,
     @SerializedName("cad")
     val cad: Double,
     @SerializedName("chf")
