+++ app/src/main/java/com/example/cryptocompose/data/model/gecko/PriceChangePercentage1yInCurrency.kt
@@ -25,7 +25,7 @@ data class PriceChangePercentage1yInCurrency(
     @SerializedName("brl")
     val brl: Double,
     @SerializedName("btc")
-    val btc: Int,
+    val btc: Double,
     @SerializedName("cad")
     val cad: Double,
     @SerializedName("chf")
